ipmi: Implement Read Event and SEL parsing
The read event mechanism is used when the BMC has an asynchronous message
for the host. It sets a flag that we read with Get Message Flags, and then
we read the message using Read Event. This event message contains a SEL
message. There are two OEM SEL messages we expect from the AMI BMC:

 - graceful power operations. This is when the BMC wants the host to
   power down or reboot.

 - PNOR Locking. When the BMC wants to access the PNOR, it requests
   that the host not touch it.

This patch implements the parsing.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>