Fix unsupported return code of OPAL_(UN)REGISTER_DUMP_REGION
Linux has been expecting OPAL_REGISTER_DUMP_REGION and
OPAL_UNREGISTER_DUMP_REGION to return OPAL_UNSUPPORTED in the event of
the call being supported on the platform (that is, the OPAL_CHECK_TOKEN
result is OPAL_TOKEN_PRESENT) but for some reason dump regions weren't
supported.

This only ever managed to be the case on P7 machines in the lab, so
it's not going to break anything that skiboot now returns the expected
result rather than OPAL_SUCCESS - which is completely the wrong thing
to return.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>