core: Add subid to load_resource()
This adds a subid to load_resource() so that sub-partitions can be accessed
inside a PNOR partition.  These sub-partitions follow the format used by the
hostboot SBE image.

The subid will match on the EC field of the SBE table of contents.  If it's
found, only that sub-partition is returned to the caller.

Current partitions (kernel and ramfs) don't support sub-partitions.  If caller
tries to access a sub-partition within these, we fail the call.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>