flash: Fix bugs for sub-partitions in flash_load_resource()
In the recent change:
  commit 73b262e768529f152ec64b4418b0c31691bc15c3
  Author: Jeremy Kerr <jk@ozlabs.org>
  core/flash: port pnor_load_resource to flash code

Some code was moved around, but it was modified introducing a number of bugs.

Firstly the size check was moved to before the sub-partition determination.
This is incorrect as the size check is against the sub-partition not the full
partition.

Secondly, it introduced a return in the error path for sub-partitions.  This
should be a goto to ensure the correct cleanup code is run.

Thirdly, subid is now set in the name lookup loop where it shouldn't be.

Fourthly, a check on the subid was removed unnecessarily.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Brown-paper-bag-acked-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>