capi: Dynamically calculate which CAPP port to use
Currently we hardware the CAPP to use PHB0.  This works for tuleta but not for
other systems.

This makes the port mapping dynamic so that first PHB to request a mode change
to CAPI will get the CAPP port mapped to it.

Calls to switch addition PHBs to CAPI mode will fail.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>