capi: Rework microcode flash download and CAPP upload
This reworks the CAPP microcode flash download and CAPP upload.

We now use load_resource() to download microcode from flash rather than
assuming we are on an FSP.  This means we can download the microcode on PNOR
based systems.

Also, currently the code associates the microcode upload with the PHB.  This
means we store one copy of the microcode for every PHB in the system.  This
patch changes this so that we only save one copy of the microcode for the whole
system.  We mark if the microcode as been uploaded in the CAPP unit based on
the chip, rather than the PHB.  We add a check in case the system has two
different chip ECs in the one system but such a Frankenmachine should never be
built!

We keep the microcode around in case we need it for a recovery event.

It also harmonises the CAPP printks to look the same.

Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>