bt: Improve error messages
Most of the error messages in the bt layer don't report enough
information to assist with debugging. This patch standardises the
information reported when the bt layer encounters an error such as a
timeout.

It also fixes a problem where expired messages would produce a
spurious warning about an incorrect netfn.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>