external/opal-prd: Add userspace support for PRD facility
This change adds an application in external/opal-prd, implementing the
userspace portion of a PRD stack.

This code is responsible for loading the HBRT code from reserved memory,
and provding hostboot runtime functionality through a set of
callbacks.

Because we may be running little-endian (and expect the HBRT code to be
big-endian), we need to thunk the endianness between calls through the
HBRT interface.

Includes multiple contributions from:

  Joel Stanley <joel@jms.id.au>
  Vaidyanathan Srinivasan <svaidy@linux.vnet.ibm.com>
  Benjamin Herrenschmidt <benh@kernel.crashing.org

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Vaidyanathan Srinivasan <svaidy@linux.vnet.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
[stewart@linux.vnet.ibm.com: squash trailing whitespace]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>