FSP/LEDS: Add device tree nodes
This patch creates a parent LED device node called 'led' under the root
'opal' device node. This also creates child device nodes under 'led'
corresponding to all individual LEDs on the system whether it is an enclosure
type or a descendant type with their location code as name. The location code
information will be used by the host to enlist and access all the individual
LEDs present on the system. The child LED device nodes also have the properties
'led-types' and 'led-loc' representing what kind of LEDs present on the same
loation code and whether it is an enclosure type LED or a descendant type LED.

Sample device tree output:
       ibm,opal {
	       ..
	       ..
              led {
                    compatible = "ibm,opal-v3-led";
                    phandle = <0x1000006b>;
                    linux,phandle = <0x1000006b>;

		    U78C9.001.RST0027-P1-C1 {
                         led-types = "identify", "fault";
                         led-loc = "descendent";
			 phandle = <0x1000006f>;
			 linux,phandle = <0x1000006f>;
		    };

		    <snip>
	      };
       };

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Anshuman Khandual <khandual@linux.vnet.ibm.com>
[stewart@linux.vnet.ibm.com: Move create_led_device_nodes to FSP platform.exit]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>