external/opal-prd: Add a helper for module loading
Rather than calling system("modprobe ...") in different locations, this
change introduces an insert_module() helper function.

We use this for the existing i2c modprobe, and add one for the ipmi
devintf module too.

We won't need the i2c-opal module, as it should be automatically
loaded by the platform device.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>