set NX crb input queues to 842 only
Set the NX crb input queues to be used by 842 only, by not reserving any input
queue slots for the sym or asym engines.  This allows the maximum number of
842 crb requests to be queued up, to (hopefully) improve 842 performance.

This patch isn't critical, as the default is no extra input queue locations
for asym, and only 1 extra input queue location for sym, so this only takes
sym's extra location for use by 842.

There is 1 each dedicated input queue location for asym and sym (and 842), so
this patch won't prevent them from working if support is ever added for them.

Signed-off-by: Dan Streetman <ddstreet@ieee.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>