mambo: Exit mambo when the simulation is stopped
When the simulation runs to completion, exit mambo. In addition to the
previous patch this allows "halt" on the Linux command line to stop the
simulation and exit mambo.

Note that this only takes effect if the "mysim go" is left to run until
the simulation is stopped. If the user interrupts the simulation, eg.
with Ctrl-C, to inspect state, then the exit has no effect.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>