external: create a GUARD partition parsing utility
This utility is designed to run completely in userland linux and make use of
OPAL APIs (exposed to userland via the linux MTD abstraction) to read, parse,
display and optionally clear firmware GUARD partition information.

The main intended use case is to have the tool run on the OpenPOWER machine
and automatically detect the flash device containing the gard data and display
information about the running system.

The tool can also read a flash imagine contained in a specified file. The tool
can also be used to parse just guard data not contained in a flash image but
rather dumped to a file.  The tool understands ECC and does its best to detect
this automatically.

There are three commands:
 - list: lists all the records
 - show: show more information about a specific record
 - clear: clear/erase a specific record or the entire flash.

The tool makes some assumptions:
 - gard data is terminated by an entry containing all FF bytes. In the case
   of the tool detecting ECC it will always read with ECC correction so the
   final all FF record will still need to contain ECC bytes.
 - The GUARD partition will not have gaps, the clear code does shuffle entires
   up and set the last one to all FF bytes.

Note: This tool has been named 'gard' because this is the traditional name for
hardware that has been detected as faulty and has been 'gard'ed. It turns out
that in the flash the partition is named 'GUARD'.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-by: Daniel Axtens <dja@axtens.net>
[stewart@linux.vnet.ibm.com: add correct (C) header to gard.c]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>