Move contents of opal.h to opal-api.h
Move the contents of opal.h, which is the definition of the OPAL API
shared with Linux, into opal-api.h.

Recreate opal.h, containing just the include of types.h as well as
including opal-api.h and opal-internal.h.

We can also remove the #ifdef SKIBOOT, because opal.h is no longer
shared with Linux.

This gets us to an opal-api.h that is purely a superset of Linux's
opal-api.h, ie. the only differences are things that are new in skiboot
and haven't yet been added to Linux.

Finally update head.S to only include opal-api.h, which is all it needs,
and avoids the need for ASSEMBLY guards in opal.h.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>