Move opal-api.h to opal-internal.h
Move the content of opal-api.h, which is currently "Internal header for
OPAL API related things" to opal-internal.h.

Recreate opal-api.h as an empty header, until the next commit, so as not
to break all the includers of it.

Include opal-internal.h from opal.h, but not for assembly users as it's
all C declarations.

Finally we need to adjust include guard names.

Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>