FSP/ELOG: Reduce elog write buffer size
Present we are allocating 64K for emergency and host buffer. But
we just need 16K (which is the max elog size). Hence reduce the
variable size to 16K.

Similarly we are allocating 256K for normal fsp write error log path.
(elog_write_to_fsp_buffer). But in reality we just need 16K. Hence
reduce this size as well.

Note that we have top level elog pool with 64 entries of 16K to hold
multiple errors. So we are fine here.

Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>