external/pflash: fix build breakage on powerpc
Because pflash is a userland tool which makes interesting use of hardware
flash accessors it needs a special version of those functions which don't
actually try to access hardware but perform various userland calls to call
into firmware to update the flash.

Sometime in the past optimisations which introduced inline functions in to the
lpc code caused it to be incompatible with the userland access functions for
powerpc.

This patch solves this by adding a special lpc header to pflash file without
these optimisations.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>