Add global CPU job queue
When we have multiple systems trying to start concurrent jobs on different
CPUs, they typically pick the first available (operating) CPU to schedule
the job on. This works fine when there's only one set of jobs or when we
want to bind jobs to specific CPUs.

When we have jobs such as asynchronously loading LIDs and scanning PHBs,
we don't care which CPUs they run on, we care more that they are not
scheduled on CPUs that have existing tasks.

This patch adds a global queue of jobs which secondary CPUs will look
at for work (if idle).

This leads to simplified callers, which just need to queue jobs to NULL
(no specific CPU) and then call a magic function that will run the
CPU job queue if we don't have secondary CPUs.

Additionally, we add a const char *name to cpu_job just to aid with
debugging.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>