Run gcc constructors
As part of setting up GCOV data structures correctly, GCC/GCOV generates
a bunch of constructor routines that the C runtime is expected to run
really early on.

skiboot was not running GCC generated constructors. Luckily, it's really
easy to do so.

This patch will run GCC constructors very early on during boot (if there
are any)

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>