Add -Tn to ibm,loc-code for ethernet adaptors
On pHyp, the location code of ethernet adaptors ends with -Tn, where n
is the port number on the card. Detect if an adaptor advertises itself
as ethernet based on it's PCI class code, and if it does, give each
PCI function a -Tn suffix, where n = function number + 1. (Therefore n
is between 1 and 8 inclusive)

This is a reasonably non-invasive change to match pHyp
behaviour. Alternatives considered were:

 - putting this in kernel space, which requires messing with the
   device tree, which is frowned upon.

 - putting this in user space, which requires patching every userspace
   tool that might consume this data, of which there are several.

As such, while it's a bit of a hack, this is the best place for it.

Signed-off-by: Daniel Axtens <dja@axtens.net>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>