flash: don't report nvram partitions through the device tree
Skiboot parses the ffs header in the nvram and writes a devicetree entry for
each partition that ffs returns. Skiboot doesn't need to do this as it exposes
access to the entirety of the nvram to linux so any user of the nvram within
linux can just read the ffs partition data themselves.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>