Have make_version be able to generate versions for tools
The goal here is to be able to set tags for versions to some of the userspace
tools that are kept in this repository.

For this to work, make_version must be able to generate a version for a tag
prefix not just the last tag in the repo.

The new usage of make_version is to specify a tag prefix when calling it so
that it knows what tag to look for. This option is ignored if not in a git
repository and the current behaviour of relying on a .version file or
$SKIBOOT_VERSION variable remains.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>