pflash: Create make dist target for pflash release tarballs
In order to be able to compile and ship pflash independently to the rest of
skiboot as it is a separate userland tool we're going to need a tarball of all
the required files to build it.

This commit adds a make dist target which copies libflash and ccan alongside
pflash sources and inserts a stripped down Makefile into a tarball.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>