libffs: Rename various offset variables to toc_offset
The word offset can be ambiguous, it can be unclear what offset the variable
refers to or what it is the offset of. As this library now has to deal with
flash with more than one libffs TOC, it makes sense to rename all uses of
'offset' to 'toc_offset' which relate to the offset of the TOC within in the
flash.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-By: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>