external/pflash: Show the partitions on the second side flash
Currently pflash is unaware that that some platforms can have two sides of
flash and therefore doesn't include the ability to print it.

The ffs TOC at the start of the flash may contain a partition called
"OTHER_SIDE" which will start with the TOC for the second side. If OTHER_SIDE
is found pflash will dump that flash information too.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-By: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>