external/pflash: Perform flashing on flash sides other than side 0
Currently pflash only looks at the first ffs TOC it finds in order to locate
an ffs partition. Problems arise when there are identical partitions in
both sides of the flash as it is current impossible to specify the second
side. This adds a parameter to specify the second side of the flash when
pflash does the lookup for the partition.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-By: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>