opal: Introduce a function to check sync/step network status.
Introduce a function to check sync/step network status for a given topology.
During chiptod initialization, check sync/step network status for backup
topology to find out whether backup topology is enabled or disabled and
update the topology info accordingly.

In the subsequent patches this function will also be used for one of the
checks to detect whether topology switch is required for TOD failure
recovery.

Signed-off-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>