Disable bust_locks for general calls of abort()
Currently bust_locks is set to true whenever abort() is called. This
was done for the lock debugging code which calls abort() as its last
step. If there is a locking error there is a good chance parts of the
console code will be deadlocked (for example if it has a lock on the
scom bus).

However busting locks can cause problems, especially now that flash
data is being read in parallel as this uses the LPC bus. Instead of
getting a nice error message showing where the code aborted it is
likely the message will go missing, or it will be garbled by the flash
data. So instead of always busting locks make it so they're only
busted if they're suspected of being broken.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>