flash: fix offset and size parameters check
Copying the flash from the host fails :

	# cat /dev/mtd0 > pnor
	cat: /dev/mtd0: Input/output error

and the kernel logs :

	[ 1357.866996] mtd mtd0: opal_flash_async_op(op=0) failed (rc -1)

It seems that the check on the parameters in the opal_flash_op() routine
are bit excessive and we fail to write or read the last block.

Here is a fix below which should be enough to catch an out of bounds
operation.

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>
Acked-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>