hw/prd: Fix handling of invlid message types
If we get a prd_message_consumed callback from the opal_msg interface,
and it has an invalid message type, then we'll end up returning with the
events_lock held.

This really shouldn't happen (as we have handlers for all message types
that we queue), but we should still handle this gracefully. This change
fixes the error path to log and error and send any further queued
messages.

Signed-off-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>