external/pflash: Add --clear command
The current pflash --erase command simply sets all the bits of the flash
back to 1 and When hostboot decides to gard everything there isn't a
useful way to get the machine booting again as even --eraseing the GUARD
partition won't leave the necessary ECC bits.

The --clear command will erase the specified partition and then fill in the
ECC bits.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Reviewed-by: Samuel Mendoza-Jonas <sam.mj@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>