dts: add helper routines to build sensor handlers
Sensors are identified using a 32-bit handler in which are embedded
a ressource class/family, a resource identifier and a resource
attribute. This patch provides a couple of routines to build core
and memory buffer (Centaur) sensors. It makes the code a little
cleaner and prepares ground for new sensor families.

Signed-off-by: Cédric Le Goater <clg@fr.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>